Revert "Refactor models into separate data package"
This reverts commit 9f47bad019da0d381fcedb8094236422e0e3c156.
