--- data/user.go
@@ -1,11 +1,0 @@
-package data
-
-import (
-	"gorm.io/gorm"
-)
-
-type User struct {
-	gorm.Model
-	Username       string
-	HashedPassword string
-}
